* Fixed incorrect type used for omp for statement.

+++ texture3d/texture3d.cpp
@@ -331,7 +331,7 @@ public:
 		const float noiseScale = static_cast<float>(rand() % 10) + 4.0f;
 
 #pragma omp parallel for
-		for (uint32_t z = 0; z < texture.depth; z++)
+		for (int32_t z = 0; z < (int32_t)texture.depth; z++)
 		{
 			for (uint32_t y = 0; y < texture.height; y++)
 			{
